1.增加了生成0-1之间随机数的函数 rand_0_1() 2.利用vector 实现和结构体一样的玩法

+++ Algrithm/GA/GA.h
@@ -5,7 +5,7 @@ bool compare(int a, int b);
 class GA
 {
 public:
-	int result;
+	float result;
 	int subscript;
 	void generate();
 };
